refactor(gb/cpu): split operand module into 2 submodules
- Operand8 and Operand16 have now their own modules contained in the Operand module
